Initial commit
Already some working PoC, as well as some tests etc.

One thing I'm somewhat unsure about is the use of panics for errors when constructing the module graph. This is mainly for simplicity/ergonomics, as there's no way to get the default operators (eg. bitor) to return Results, so we have to panic there anyways, and then it makes sense to consider panicking for all module graph construction errors for consistency, which actually probably makes sense anyways, as it's typically a programmer error that causes this, not a user error. Additionally, this preserves file/line information when the error happens in the backtrace without having to pull in external error handling dependencies just for that. This is quite important, as the whole point of erroring during graph creation is to make sure the error is caught as close to the place that caused the error as possible, to make debugging RTL code written in this HDL as easy as possible.

Anyways, a bit rambly, but this is working, yay