Reorganization, started on docs, more impls, etc
 One thing that's still up in the air here is how to make conditionals etc easier to do, instead of explicitly using a bunch of mux's. So far I'm experimenting with a basic declarative macro that attempts to do what I had in python before, and it appears to mostly work, but ends up with more expansion then I would like, and doesn't correctly parse all combinations of if's and assignments.

Ideally, what we'd like to do is allow assignments to do what they normally do, but then patch up the results of the assignments with the selectors from the if statements, like we do in python. However, this is proving difficult with regular macros, since we can't have a set of the newly-assigned variables that we can use to patch at the end of each if statement. However, this might be straightforward to do with a procedural macro, assuming it would still be possible to support a similar (or better) syntax, so maybe that's worth looking into.