Proper literals, registers, cleanups
Lots of TODO's in here and stuff I'm not totally sure about, but it's really cool to see this working in sim, and without too much effort. There's also a couple unsafe blocks in the sim generation code, but it's not many, and they're used in trivial cases where we know the lifetimes work out, so I don't think it's a big deal.