Small fixes
One is just a TODO; the other is to use a HashMap instead of a BTreeMap in the sim codegen. I typically like to use BTreeMap's for compiler-y things like this because it preserves node ordering between runs which is nice for debugging/testing, but in this case, since we're storing pointers, we already lose that property to begin with. In that case, using a HashMap should be faster, and more importantly, communicates that we don't have any reason to order the elements in the map.