Implement/document/test Signal::bits
This one was a bit more complicated than bit, since it's possible to end up requiring a cast to a type other than a boolean. The code I have so far is probably robust enough, but we're going to want to come up with and/or generate some more complicated test cases that cover a wide range of combinations of features.