Implement/document/test Signal::repeat
I'm really not liking how this widening/narrowing is looking in the sim gen code. This probably isn't a huge deal since it's all tested and will be very easy to handle in the verilog gen code, but I suspect a _lot_ of what I just wrote for repeat will also apply to concat. In any case, there's surely some latent simplifications here.