Refactor sim gen code
This makes it a bit longer, but conceptually simpler, at least in my head. The main idea is to introduce another IR to represent the assignments/expressions in the prop fn, which turns gen_expr into compile_signal, where it can operate in a more semantically-rich domain, and do things like 'generate a cast' and 'mask off these bits' without working with strings directly. Once we compile all the assignments for the prop fn, the assignments are then serialized via the code writer (which is now trivial).

Not only does this make the code semantically much clearer by separating compilation and string generation, it also allows us to more easily introduce temp variables, which I've decided to do MUCH more liberally. While this makes the output code more verbose in very basic cases, it scales orders of magnitude better when, for example, long expansions are used (eg. high().repeat(128).bit(27).repeat(128).bit(0).repeat(128)). Since it's surprisingly easy to create pathological cases like this, it's important that the compiler doesn't blow up when it happens.

There are still a couple TODO's, but overall, I'm MUCH happier with this, and trust it a lot more.