Initial instantiation impl
Has a couple tests, but no error tests or documentation, so this is certainly incomplete. But it was well worth getting a happy path working first; I already went through a couple ways to implement this before I settled on something I'm happy with, so that's good. One thing that could certainly be improved though is how we specify connections to instantiated modules (both inputs and outputs). I think this should really only be a matter of introducing the right syntactic sugar and technically we don't actually need anything as long as it works, but I know from experience that verbose input/output connection code is error-prone and annoying to deal with, so I'd like to have at least some kind of offering there.