Document/test Instance::drive_input
One thing not covered here is trying to detect combinational loops, which I'm not totally sure is possible/appropriate to do here. I'd like to not defer this until codegen but we may not have all of the information we need to do here, and it may not make sense anyways since it would theoretically be less efficient to check for this whenever an input is connected instead of once when code is being generated. Need to think more about that one.