Hide code_writer module and remove custom error types
Currently the only error that's not a panic is an indentation underflow error in the CodeWriter, which represents an internal compiler error anyways, so could just as well be a panic itself. If we exclude that then we don't have to expose this error or the code_writer module at all, which is good, as that wasn't meant to be a public API anyways, so we just end up removing code.