Add From conversions from unsigned integral types to Value
This is a bit unergonomic in most cases since unconstrained literals default to i32, which isn't currently implemented, as we don't have anything that handles signed values currently, so they wouldn't make any sense. This means all numeric literals must be suffixed with the intended type which hurts readability for many common values, although it does make these literals very consistent. In any case, this is definitely better than the verbose explicit construction we had before, and it looks great for boolean values.

Note also that this doesn't cover default reg values, since None doesn't specify enough type information to work properly, so I'm thinking instead I might add a different method for Registers that supplies a default value, which works around the issue and makes the common case (don't care about default value) less verbose. Another way we could do this is to make default values mandatory which might be nice, although a common optimization is to not initialize registers in the datapath in order to reduce combinational latency, and it's nice to be able to support this (although we should probably support ways to randomly initialize these values in sim).