Implement/document/test Add for Signal
I spent quite a long time thinking about the API for this. Technically, a sum of two numbers with n and m bits respectively should yield a result with max(n, m) + 1 bits. In the original python kaze, this is how I decided to implement it, but I found that most of the time, I didn't want the carry bit, and ended up having to explicitly truncate the result quite often. So, I've decided that the output should be truncated by default, and the inputs can be appended with a 0 bit if carry is desired. I'm also considering adding an overflowing_add fn for Signal that, similar to overflowing_add for integer types, would yield a tuple containing the truncated result as well as a single-bit carry result, if this is something that happens often. A similar thing could be done for Sub.

The other thing about this API is that it requires that lhs/shr have the same bit width. This is not strictly necessary, but I figured it would smoke out potential bit width errors in user code. I may be wrong about this, but I can always relax this error case without breaking any user code, whereas the opposite would be a breaking change. This also makes the operator symmetric with other existing binops, so I think this is a good choice.