Change register API to accept default values in another fn
This is mostly for ergonomics, but also forces the internal model to allow registers without default values. It's often the case that a datapath reg doesn't want a default value in order to reduce combinational logic, so this is something we want to allow. Tests have been adjusted to account for this (although the required sim change(s) were minimal).