Refactor InstanceStack into ModuleContext
This reduces tuples of (refcounted linked list, node pointer) to (context pointer, node pointer), which feels a bit better. Had to get explicit with some lifetimes, but it ends up working out in the end.

I'd still like to perhaps separate the register gathering/validation pass into a separate pass so that both sim gen and verilog gen can use the same code, but I'm not 100% sure what that will look like. In either case, I think this is a step in the right direction.