Require names for instances
This is primarily for error reporting which isn't implemented yet. I'm considering the same for regs.

Currently only public internally in the crate, and unused, so there's a warning generated, but this is temporary. There's also no tests that this name is unique or non-empty (similar to inputs/outputs/etc), and the docs could probably be better, but let's start with this for now.