Add indirect module hierarchy loop detection
Slightly more complicated, but not really. Technically only one kind of error would suffice here, but I think it's nice to be able to clarify to the user whether a recursive definition is formed directly (i.e. instantiating module a inside itself) or indirectly (via some other instantiations in between). In the future, it might make sense to print out the path between the module and its instantiation, which we should have enough information for, but I can't be bothered to deal with that now.