Implement/test error check for all undriven inputs on all instances
We want to be conservative here and cover the whole hierarchy to avoid surprising errors that could arise if an undriven input were not reachable, then code was generated successfully, and then a change was made that indirectly made that input reachable, and an error would be generated that would probably seem unrelated to the actual change. I also like the idea of always driving every input, since it forces the user to make the decision whether to connect an input to something meaningful or just tie it high/low.