Require names for registers
Similar to with instances, this is primarily for error reporting which isn't implemented yet. We can also decorate the generated sim code with the names, though, which is helpful for debugging at least. Also missing tests to make sure the name is unique and non-empty, and still no reg docs.