Make generated rust simulator structs have new constructor
This follows a note in C-COMMON-TRAITS about this ctor always being expected, even if its implementation is equivalent to default (which is how it's implemented here). Note that the rest of C-COMMON-TRAITS isn't necessarily respected in this version of the library as most public types don't implement most traits. I think this is fine, as I'm not entirely sure how to handle these yet, and I'd rather not expose implementations that I know may change until I figure it out.