Implement/document/test signed comparison ops
This includes lt_signed, le_signed, gt_signed, and ge_signed. This is a bit of an unfinished story since we still don't support signed literals, but I'm thinking that if we keep following this pattern with the low-level stuff where we keep everything unsigned internally until the few cases where we have to do a particular signed op, then we might solve that by allowing signed literals, and just having more complex bounds checks to cope with that. It should still totally be the case that higher-level abstractions (eg signed/unsigned integer wrappers) could be built on top of this API, so I think we're still on the right track.

Part of this impl that I'm not 100% happy with is that I extended ValueType to include signed types, and then added unreachable match arms for them in a few places in the sim compiler. It may be cleaner to have separate signed/unsigned value type types, but then bool seems to be somewhat of a problematic case. Either way it's not a huge deal since we have good test coverage and the problematic cases should be caught at the API-level (hence the use of unreachable in the first place). but I may take another look at it.

Also fix some small typos in earlier docs.