Replace sugar macro with if/else_if/else construct
This is still marked as unstable, as I'm not sure it's what we want long term, but my test projects look quite nice with it, and it generates nice code, so let's continue going down this path. Unfortunately, rustfmt doesn't seem to like the added scopes and makes them look weird, which is really a shame.