Implement Eq/PartialEq/Hash for Signal and ModuleContext
Primarily, this allows us to remove the last remaining unsafe block in the entire repo, and has the nice benefit that now these are implemented for public types (at least the references, which is all the API actually exposes).

These impl's are missing targeted tests and documentation, as I haven't decided if/how/where I should actually test/document them specifically yet. Since they're part of the public API, tests should probably be added. I think I'll ponder it a bit more and address that in a later patch.

If we _do_ want to document these, then we should make sure to mention that `==` for two `Signal` references does _not_ return a new signal that can be used in conditionals etc. The type system should already catch this, though.