Refactor binops/sim ir
The main idea is that we have some add/sub- and comparison-specific code, and I think it's a better idea to separate these ideas in the type system instead of having more conditional logic just because the shapes of the data structure are similar. It may be that I end up merging these again later and separating only the op types, since there's still some commonality in the the handling of these variants, but we'll see. For now, I want to see where this goes as I implement shifts, so we'll see.

The sim ir is changed mainly to differentiate infix binops (which use to just be binops) from unary member calls, again to remove extra conditional logic, as these are separate constructs in the target language.