Implement/document/test Shl for Signal
This ends up being relatively complex in the compiler since rhs can be any valid number of bits, so we need to make sure we generate code that properly handles overflow. I considered instead always limiting the bit width of the rhs signal, however this doesn't work cleanly with bit widths for self that aren't powers of two.