Implement/document/test Mem construct
This was quite big as there are a lot of considerations here, and I 100% have not covered all possible use cases, but I think I've arrived at something I'm reasonably happy with.

Also included is a couple new TODO's for things I think could be improved in other parts of the code, and some more sim refactorings which mostly refactor assignments for reset and posedge_clk to use the same Assignment structure as prop assignments, which is nice. There are some implementation TODO's I'd like to revisit later though.