Separate StateElements from sim Compiler
Since these don't need to be mutable past the initial gather stage, we can separate it into its own object, which means we no longer have to clone it on the top-level, and it makes the code a bit cleaner. Additionally, this should make it more natural to add other state elements (for example memories).