Support module instances in SystemVerilog output
Also loads of refactorings/reorganizations in the verilog compiler. The biggest major change is that we no longer collect all state elements as we do in the sim compiler (the state_elements module has been moved back to the sim module etc) and instead only look at things local to the current module (after validation ofc). This means we're not stripping any unused stuff, but I think that'll be OK. We also make temp nodes for all outputs of all instances even if they're unused; this one might actually cause some warnings in compilers further down the line, but I'll have to see it happen before doing anything about it.

This should also make it easier to support memories later.