Change Mem semantics for simultaneous read/write
When both a read and a write are asserted for the same location in memory on the same cycle, the read will now return the _previous_ value at the memory location, _not_ the newly-written value. This was done to work around behavioral differences observed in synthesis tools. Particularly, I was able to find a case where a certain FPGA vendor's synthesizer would decide to infer block RAM or distributed RAM depending on the capacity required for the memory (all else being equal), and this actually yielded different behavior on the device. It seems these new semantics are less ambiguous and generally easier to implement.