Add additional if_ sugar variants for tuples with up to 12 elements
Previously we had 8 which should be more than enough, but I ended up finding a case with 9 so I bumped it up. Naturally I think this is a sign that something's not being done in a great way, but that doesn't mean we shouldn't support it anyways, as in this case it's quite helpful to do the slightly wrong thing first, get it working, and then refactor later.

I still have yet to figure out a nice way to generate these with macros so I continued the current pattern for now.