Implement/doc/test Signal::mul
This represents unsigned multiplication. Given inputs of m and n bits respectively, it will produce a product of m + n bits.

I'm not entirely sure if I want to support signed multiplication out-of-the-box or not, and until I figure that out, it won't be part of the API (and I can always change unsigned mul's API to find something more symmetric if I have to later).