Implement/doc/test Signal::mul_signed
Like Signal::mul, this represents multiplication, but it treats its arguments as signed values instead of unsigned, and produces a potentially-signed result.

While this certainly works well, the story feels a bit incomplete, as kaze still has no way to create literals from signed numbers, so that will likely have to be added soon as well, although it's not technically necessary since those literals can be expressed as unsigned values with the same bitwise representation currently.