Don't derive `Default` trait to help generate module sim initialization code
Before we implemented `Mem` it was sensible, but now it's required to use new instead of default when a Module contains a Mem instance. This means that the derived Default impl isn't actually valid anymore, so we should remove it, which requires us to explicitly initialize all of the struct's members, not just the `Mem` elements.

Fixes #3.