Eliminate recursive graph traversal
In a few places, we were recursively traversing the module graph. While many graphs are quite small, some are practically too large for this, and we end up blowing the stack.

To fix this, I've converted some recursive implementations to iterative ones in the compiler. This is a bit more complicated, but necessary to be able to handle the use cases we want to handle.

A surprising case that I also found was Signal::bit_width, which could blow the stack for some expressions not terminated by state elements. To fix this, I've added bit_width fields to many of the SignalData variants, so we calculate that up front instead of recursing.

Also fix some small format warnings.

Fixes #11.