Fix stack overflow bugs when cloning sim::ir::Expr
This seems to have slipped by in the recent changes, but totally makes sense - the previous definition for these Expr's used Box<Expr> for children, and the default clone impl would recursively traverse the Expr tree. Not only can this blow the stack (which I didn't notice on mac but was obvious on windows), it also wastes a lot of extra time/mem when performing these clones. This isn't necessary at all, since we can just use references and alloc Expr's in another arena instead, which solves all of these issues.