Implement/doc/test Signal::reg_next and Signal::reg_next_with_default
These are convenience methods to produce a new Signal whose value follows this Signal's value, but delayed by a cycle, and optionally having a default value. This is a very common pattern in RTL, and I've been using it extensively in my test project(s) without any changes to the pattern - so I thought it was appropriate to promote it to a proper API feature.