Fix duplicate trace member names in rust sim gen
This was pretty easy to identify/fix - I'm not sure how I made this mistake originally, but I'm guessing that `trace_signals` was originally a `Vec`, not a nested `Vec` in a `HashMap`, which would have meant that `trace_signals.len()` would have correctly enumerated all trace signals. With a nested structure, this isn't the case, and we have to track this explicitly instead.