The great module/instance collapse
I want to move kaze to a language where modules are expressed as Rust structs, and instances are expressed as instances of those Rust structs. While the exact manifestation of this is in flux and still fairly unclear to me, I've taken some (relatively drastic) steps here to work towards that goal. These changes are far-reaching, so there's little point in trying to break them up into smaller changes. Additionally, not all of this work is done; not all tests pass, some things are missing documentation, and we need some proper error messages. Still, I need to get this on master to be able to move forward with this and other work. We have to start somewhere, so here goes.

Some major points:
 - Remove `Instance`
 - Introduce dedicated `Input`/`Output` structs, which are now constructed directly by a `Module`
 - Introduce `ModuleParent` trait, which is implemented by `Module` and `Context` and provides a `module` function, which is used to construct a `Module`. This allows `Module`s to form a hierarchy like we want.
 - Make `Signal` a trait, and split its implementation between abstract stuff that only deals with `Signal`s, and more concrete implementation details, hidden in the new `InternalSignal` struct. Implement `Signal` for `InternalSignal`, `Input`, `Output`, and `Register`, which now means that a user doesn't have to use `Register::value` to get a `Signal` for the register's current value; they can use the `Register` instance in-place.
 - Users now typically refer not to `Signal` directly, but as it's now a trait, `dyn Signal`.
 - Temporarily remove verilog generation (obviously this needs to return, but some dust needs to settle first)

And lots of other, smaller bits/bobs.